requirements.txt
crewai
langchain
openai
django==5.1.7
djangorestframework
django-cors-headers
